requirements.txt
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
ffmpeg-python>=0.2.0
imagehash>=4.3.1
numpy>=1.24.0
pillow>=10.0.0
pandas>=2.0.0
tqdm>=4.65.0
flask>=2.3.0
werkzeug>=2.3.0

